+++ MyEssentialFeed/MyEssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -58,25 +58,3 @@ private final class FeedViewAdapter: FeedView {
         }
     }
 }
-
-private final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
-    private let feedLoader: FeedLoader
-    var presenter: FeedPresenter?
-    
-    init(feedLoader: FeedLoader) {
-        self.feedLoader = feedLoader
-    }
-    
-    func didRequestFeedRefresh() {
-        presenter?.didStartLoadingFeed()
-        
-        feedLoader.load { [weak self] result in
-            switch result {
-                case .success(let feed):
-                    self?.presenter?.didFinishLoadingFeed(with: feed)
-                case .failure(let error):
-                    self?.presenter?.didFinishLoadingFeed(with: error)
-            }
-        }
-    }
-}
